This is the complete CWRU bash 1.13 patchlevel 2 source distribution as suggested by Chet Ramey.
The only source change other than the makefiles is the addition of
'#define HAVE_UID_T' in machines.h.  The patches were sent to Chet.

This version adds the 'ulimit -u #' which enables one to increase the
per-user process limit above 40 and build X.  The bash binary builds
with a simple 'make' in the top level directory.