This is a patch that transforms trn 3.3 into trn 3.4.
This release supports XTHREAD, XOVER, XINDEX, LISTGROUP, faster XHDR
response from overview files, etc.

Once the source is patched you should re-run Configure, just to be
safe:

	Configure -r -e -d -s

should be reasonably fast.

Finally do a "make" and a "make install" and you're done.

See the file "NEW" for a list of what's new in trn 3.4.

If you plan to make trn on multiple machines see the Policy.sh file in
the hints subdirectory -- you can use this file to make configuring an
additional system or the mthreads package easier by putting it in the
hints directory of the freshly unpacked package and specifying it when
you specify the other hint files.

Wayne Davison
davison@borland.com